remove lots of duplicated code for gdrive and gsheet
